chore: move bookmark file
