+++ apps/web/src/common/components/bookmark/BookmarkDeleteModal.vue
@@ -8,6 +8,7 @@ import { i18n } from '@/translations';
 
 import { assetUrlConverter } from '@/lib/helper/asset-helper';
 
+
 import { gray } from '@/styles/colors';
 
 import { BOOKMARK_MODAL_TYPE } from '@/services/workspace-home/constants/workspace-home-constant';
