+++ apps/web/src/services/preference/routes/admin/routes.ts
@@ -28,7 +28,6 @@ const adminPreferenceRoutes: RouteConfig = {
             path: 'workspaces',
             name: makeAdminRouteName(PREFERENCE_ROUTE.WORKSPACES._NAME),
             meta: {
-                lsbVisible: true,
                 menuId: MENU_ID.WORKSPACES,
                 translationId: MENU_INFO_MAP[MENU_ID.WORKSPACES].translationId,
             },
@@ -48,7 +47,6 @@ const adminPreferenceRoutes: RouteConfig = {
             path: 'domain-settings',
             name: makeAdminRouteName(PREFERENCE_ROUTE.DOMAIN_SETTINGS._NAME),
             meta: {
-                lsbVisible: true,
                 menuId: MENU_ID.DOMAIN_SETTINGS,
                 translationId: MENU_INFO_MAP[MENU_ID.DOMAIN_SETTINGS].translationId,
             },
