+++ apps/web/src/services/workspace-home/store/bookmark-store.ts
@@ -18,9 +18,9 @@ import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-worksp
 
 import getRandomId from '@/lib/random-id-generator';
 
+import { fetchFavicon } from '@/common/components/bookmark/composables/use-bookmark';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
-import { fetchFavicon } from '@/services/workspace-home/composables/use-bookmark';
 import { DEFAULT_BOOKMARK } from '@/services/workspace-home/constants/workspace-home-constant';
 import type { BookmarkItem, BookmarkModalType, BookmarkModalStateType } from '@/services/workspace-home/types/workspace-home-type';
